modern: reorganise routes
