+++ routes/map.js
@@ -1,0 +1,8 @@
+import { Router } from 'express'
+const mapRouter = Router()
+
+mapRouter.get('/', (req, res) => {
+	res.render('map', { session: req.session })
+})
+
+export { mapRouter }
